Fixed remove command
The remove command did not delete the main config file
only the pilot link and the config .d directory. In addition
the removing code was not strict enough on checking if the
provided app path is really something that should be deleted
by oci-ctl. Thus I also added sanity checks to prevent
accidental deletion of arbitrary files on the system