Fixed sci setup and command execution
sci was a bit too picky on mounting proc, sys and run.
If called from within an initrd it could happen that these
mountpoints already exists. I turned this condition into
a warning rather an error. Furthermore sci executes itself
when the idea was that sci executes a given command. I
refactored the code to allow calling a given command and
also let the firecracker instance shut down after the
command execution. The code also leaves two TODO code
points which I'd like to fix later and together with the
team