Fix for commands with long output
Added interactive boot argument, when set to false(default value)
sci will grab the whole output of the command and wait until all
data from the console will be read to provide to the user. The
interactive option shall be set to true for all applications where
we don't want to grab the output such as bash or python interpreter etc.
This solves #82