Support pulling image from components
So far pulling a firecracker image was limited to the KIWI
kis image type. With this commit it's also possible to pull
an image from it's components; rootfs, kernel and optional
initrd. This allows for example to make use of the firecracker
hosted CI images or any other rootfs image in combination
with some kernel/initrd. During the pull of such an image
there is also a check for the sci binary inside of the rootfs.
If it does not exist it gets added to the rootfs such that
the firecracker-pilot can work with it. This Fixes #99