Make sure to create /overlayroot
When using arbitrary rootfs images it might happen that there
is no /overlayroot mountpoint which we use in the pilot to
setup an overlay. In addition search for tools like switch_root
in the former location /sbin to stay backward compatible with
linux versions prior the /usr/sbin move. Newer linux versions
still maintains symlinks from /sbin to /usr/sbin for compat
reasons.