Added instance resume process to firecracker pilot
When using a firecracker instance in resume mode the first
call has to run firecracker in the background, open a connection
to the vsock and send the command. Any subsequent call only
sends commands to the existing instance. If the instance gets
deleted the pilot has to check if the VM is still there and
behave properly. Meaning vmid (PID) and uds socket file must
be cleaned up prior launching a new instance. The commit here
implements the processing of this workflow. It does not
implement the actual connection to the instance and sending
of commands. This will follow. As of now only debug messages
telling what should happen are printed. This is related
to Issue #87