Increase max retries on resume type VM
When connecting to a resume type VM for the first time, the
wait time was not long enough. Increasing the retry count
will solve this issue. If after this delay the VM is still
not responsive on the vsock it's likely an issue on the
VM startup itself and ok to close. This Fixes #122