Regression and partial revert from the clippy set
The changes done to make clippy happy also changed the logic
of the code at several places. This was exactly the reason
why I wanted small reviewable pull requests. The commit here
is a partial revert of the code prior the big clippy changes.
The reason for the changes here is because the template
data was was no longer imported and also options given on the
commandline were no longer added to the flake yaml file.
I reverted to the version before clippy and manually fixed
all clippy issues while testing the code to be still functional.
This is the result commit of this action.