Refactor port allocation for UDS socket
When using the firecracker-pilot in a resume type flake, the pilot
creates a UDS socket for listening on data. Each of this socket
connections requires a free port. The former implementation selects
the socket by a random number within a given range. However, this
is not a stable solution and the risk to select an already occupied
port is high. This commit refactors the port allocation to
FIRECRACKER_VSOCK_PORT_START + the PID of the pilot process.

In addition the commit adds to allow to read runtime options which
are only meant to be used for the pilot and should not interfere with
the standard arguments passed along to the command call. For this
purpose we deviate from the standard Unix/Linux commandline format
and treat options passed as %name:value to be a pilot option.
This commit adds the handling of the followng pilot option:

%port:number
    Allows to specify a custom UDS port used for the
    VSOCK based communication channel in a firecracker flake
    configured for resume mode

This Fixes #114